refactor: 클라이언트로부터 콘솔 입출력 시에 사용되는 디렉토리 생성
