+++ src/main/java/org/example/Main.java
@@ -1,5 +1,7 @@
 package org.example;
 
+import org.example.app.Client;
+
 public class Main {
     public static void main(String[] args) {
         Client client = new Client();
